removal of public and restructure of files
