+++ .gitignore
@@ -39,4 +39,3 @@ jspm_packages
 # DB Config File
 dbconfig.json
 
-/public
