+++ app/js/models/quote.js
