+++ app/js/services/quotesService.js
