--- public/index.js
@@ -1,11 +1,0 @@
-'use strict';
-
-var _server = require('./config/init/server.js');
-
-var _server2 = _interopRequireDefault(_server);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-var server = new _server2.default();
-
-server.start();
